containers/app/streamlit/requirements.txt
numpy==1.23.5
Pillow~=9.2.0
streamlit==1.15.2
plotly==5.11.0
pandas==1.5.2
tensorflow 
requests
